requirements.deploy.txt

PyYaml==6.0.2
setuptools==75.1.0
tox==4.20.0
twine==5.1.1
